refactor: rename core to domain and app to application
