+++ src/domain/window-merge.ts
