+++ src/ports/tab-group.ts
@@ -1,4 +1,4 @@
-import type { GroupId, MoveToWindow } from '../core/window-merge.types';
+import type { GroupId, MoveToWindow } from '../domain/window-merge.types';
 
 export type TabGroupPort = {
 	readonly moveGroup: (groupId: GroupId, moveProperties: MoveToWindow) => Promise<void>;
