+++ src/ports/window.ts
@@ -1,4 +1,4 @@
-import type { WindowSnapshot } from '../core/window-merge.types';
+import type { WindowSnapshot } from '../domain/window-merge.types';
 
 export type WindowPort = {
 	readonly getAllWindows: (populate: boolean) => Promise<readonly WindowSnapshot[]>;
